feat: combat, stats, levelling, and inventory (#76)
* feat: basic skeleton groundwork for combat

* feat: gracefully logout without causing server errors

* fix: display correct monster name on look command

* feat: initial working combat systems

* fix: enable and fix clippy::cargo warnings

* refactor: simplify position and velocity into single struct

* refactor: make client naming convention distinct from server

* fix: crash when player enters top row

* refactor: first pass at clippy pedantic warnings for integer casts

* refactor: first pass at clippy pedantic warnings for client and lib

* refactor: first pass at clippy pedantic warnings for server

* refactor: second pass at clippy pedantic warnings

* refactor: third pass at clippy pedantic warnings

* refactor: resolve final clippy pedantic warnings

* feat: make player HP updates visible to player after combat

* feat: add uuid to player_details

* refactor: use f32 stats more accurate combat calculations

* fix: revert incorrect player str default

* refactor: split systems into separate modules

* feat: implement player death mechanics

* refactor: remove the need for player_position_updates resource

* chore: remove todo comments

* feat: implement monster deaths

* feat: implement sending combat system messages to players

* feat: implement sending combat system messages to players

* docs: update readme to reflect combat feature

* feat: implement monsters attacking players

* feat: implement monster respawn mechanics

* fix: skip combat checks if either attacker/defender already dead

* feat: implement health regen every second server tick

* feat: make player attacks only happen on server tick

* refactor: change look command to l key

* docs: clarify controls

* chore: clarify attack miss message

* refactor: move exp and lvl to stats

* feat: implement exp gain on killing enemy

* fix: only give exp on monster deaths, not players

* feat: implement levelling up and stat gains

* chore: change clippy lint checks

* fix: new max hp should be rounded

* feat: implement inventory system structs

* feat: implement inventory system

* fix: make resizing the client window more robust

* feat: dynamically resize viewport based on window resize

* docs: clarified comment

* feat: gain gold from killing monsters

* refactor: tweak side status bar layout

* refactor: create network message system

* refactor: tweak side status bar text

* fix: gain gold from multiple monsters

* feat: added trinket item type

* refactor: move entity update messages to broadcast system

* fix: player walking on item

* refactor: minimise broadcast size to minimum required

* fix: remove dead monsters from player display

* fix: send entity broadcasts individually so max packet size is not exceeded

* feat: add item ids

* fix: look underneath now works

* docs: added CHANGELOG.md

* docs: add more details to README.md

* chore: update dependencies

* feat: implement pickup command with some bugs

* docs: updated CHANGELOG.md

* fix: duplicate items being added to world

* fix: item disappears correctly from client viewport when picked up

* chore: add local startup script to .gitignore

* feat: new default maps and spawns

* feat: add drop item functionality

* docs: clarified description

* docs: clarified description

* docs: clarified description

* docs: clarified description

+++ rustyhack_lib/src/background_map.rs
@@ -11,12 +11,14 @@ pub struct BackgroundMap {
 }
 
 impl BackgroundMap {
+    #[must_use]
     pub fn data(&self) -> &Vec<Vec<Tile>> {
         &self.data
     }
 
-    pub fn get_tile_at(&self, x: usize, y: usize) -> Tile {
-        self.data[y][x]
+    #[must_use]
+    pub fn get_tile_at(&self, x: u32, y: u32) -> Tile {
+        self.data[y as usize][x as usize]
     }
 }
